+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v0.3.0
+* Add ability to save a request to file
 * Split cmd and lib
 
 ## v0.2.1 (2017-06-04)